+++ ChangeLog.md
@@ -12,6 +12,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - :openai module support for audio speech generate
+- :openai module support for audio speech transcription