+++ ChangeLog.md
@@ -12,7 +12,8 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - :openai module support for audio speech generate
-- :openai module support for audio speech transcription
+- :openai module support for audio speech transcribe
+- :openai module support for audio speech translate