+++ ChangeLog.md
@@ -14,7 +14,12 @@ All notable changes to this project will be documented in this file.
 - :openai module support for audio speech generate
 - :openai module support for audio speech transcribe
 - :openai module support for audio speech translate
-- openai vision examples
+- :openai module support for file upload
+- :openai module support for file list
+- :openai module support for file retrieve
+- :openai module support for file delete
+- :openai module support for file retrieve content
+- OpenAI vision examples